+++ decide/authentication/forms.py
@@ -1,5 +1,4 @@
 from django import forms
-from django.contrib.auth.models import Group
 from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth.models import User
 
@@ -12,13 +11,3 @@ class SignUpForm(UserCreationForm):
     class Meta:
         model = User
         fields = ('username', 'first_name', 'last_name', 'email', 'password1', 'password2')
-
-class importForm(forms.Form):
-    name = forms.CharField(max_length=80, min_length=1, label='Nombre del grupo', required=True)
-    is_public = forms.BooleanField(label='Público', required=False)
-    file = forms.FileField(label='Archivo txt o xlsx', required=True)
-
-
-class exportForm(forms.Form):
-    group = forms.ModelChoiceField(label='Selecciona grupo a exportar', queryset=Group.objects.all())
-
