+++ decide/authentication/templates/master.html
@@ -31,8 +31,8 @@
             </li>
 
 
-            <li><a href='/authentication/groups/import'>Importar grupo desde Excel.</a></li>
-            <li><a href='/authentication/groups/export'>Exportar miembros de un grupo a Excel.</a></li>
+            <li><a href='/census/groups/import'>Importar grupo desde Excel.</a></li>
+            <li><a href='/census/groups/export'>Exportar miembros de un grupo a Excel.</a></li>
 
             {% if user.is_superuser%}
             <li><a href="/admin">Zona de administración</a></li>
