+++ decide/authentication/urls.py
@@ -12,7 +12,5 @@ urlpatterns = [
     path('logout/', LogoutView.as_view()),
     path('getuser/', GetUserView.as_view()),
     path('register/', RegisterView.as_view()),
-    path('groups/import/', importGroup),
-    path('groups/export/', exportGroup),
     path('activate/<id>/<token>/', activate, name='activate'), 
 ]
