+++ decide/base/tests.py
@@ -46,7 +46,7 @@ class SeleniumBaseTestCase(StaticLiveServerTestCase):
         self.base.setUp()
 
         options = webdriver.ChromeOptions()
-        options.headless = False
+        options.headless = True
         options.add_argument("--incognito")
         self.driver = webdriver.Chrome(options=options)
